Restructure command call syntax due to review
