+++ azure_cli/app.py
@@ -25,6 +25,7 @@ class App:
     def __init__(self):
         app = CliTask()
         action = app.cli.get_command()
-        task_class_name = action.title() + 'Task'
+        service = app.cli.get_servicename()
+        task_class_name = service.title() + action.title() + 'Task'
         task_class = app.task.__dict__[task_class_name]
         task_class().process()
